# backend/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Global Multi-Entity ERP API",
    "version": "0.4.0",
    "description": "API contract for global multi-entity accounting endpoints under /api/v1.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.global-ledger.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/org/tree": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "getOrgTree",
        "summary": "Get organization tree",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier; optional if available in JWT",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization tree",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgTreeResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/group-companies": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listGroupCompanies",
        "summary": "List group companies",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group company list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Org"
        ],
        "operationId": "upsertGroupCompany",
        "summary": "Create or update group company",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupCompanyInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group company created or updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCompanyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/legal-entities": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listLegalEntities",
        "summary": "List legal entities",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "groupCompanyId",
            "required": false,
            "description": "Group company identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "countryId",
            "required": false,
            "description": "Country identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Legal entity list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Org"
        ],
        "operationId": "upsertLegalEntity",
        "summary": "Create or update legal entity",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalEntityInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CreatedResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/operating-units": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listOperatingUnits",
        "summary": "List operating units",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "legalEntityId",
            "required": false,
            "description": "Legal entity identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operating unit list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Org"
        ],
        "operationId": "upsertOperatingUnit",
        "summary": "Create or update operating unit",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperatingUnitInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CreatedResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/countries": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listScopedCountries",
        "summary": "List countries visible in scope",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Country list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/fiscal-calendars": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listFiscalCalendars",
        "summary": "List fiscal calendars",
        "parameters": [
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiscal calendars",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": [
          "Org"
        ],
        "operationId": "upsertFiscalCalendar",
        "summary": "Create or update fiscal calendar",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiscalCalendarInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/CreatedResponse"
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/fiscal-calendars/{calendarId}/periods": {
      "get": {
        "tags": [
          "Org"
        ],
        "operationId": "listFiscalPeriods",
        "summary": "List fiscal periods for a calendar",
        "parameters": [
          {
            "in": "path",
            "name": "calendarId",
            "required": true,
            "description": "Fiscal calendar identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "tenantId",
            "required": false,
            "description": "Tenant identifier",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "fiscalYear",
            "required": false,
            "description": "Fiscal year",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fiscal periods",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnyObject"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/org/fiscal-periods/generate": {
      "post": {
        "tags": [
          "Org"
        ],
        "operationId": "generateFiscalPeriods",
        "summary": "Generate fiscal periods",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiscalPeriodGenerateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fiscal periods generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiscalPeriodGenerateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "401": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "CreatedResponse": {
        "description": "Created",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Created"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "AnyObject": {
        "type": "object",
        "additionalProperties": true
      },
...
      "OrgTreeResponse": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "countries": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "legalEntities": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "operatingUnits": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": [
          "tenantId",
          "groups",
          "countries",
          "legalEntities",
          "operatingUnits"
        ]
      },
...
      "GroupCompanyInput": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "code",
          "name"
        ]
      },
      "GroupCompanyResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "nullable": true
          },
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "ok",
          "tenantId",
          "code",
          "name"
        ]
      },
      "LegalEntityInput": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "groupCompanyId": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "taxId": {
            "type": "string",
            "nullable": true
          },
          "countryId": {
            "type": "integer",
            "minimum": 1
          },
          "functionalCurrencyCode": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          }
        },
        "required": [
          "groupCompanyId",
          "code",
          "name",
          "countryId",
          "functionalCurrencyCode"
        ]
      },
      "OperatingUnitInput": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "integer",
            "minimum": 1
          },
          "legalEntityId": {
            "type": "integer",
            "minimum": 1
          },
          "code": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "unitType": {
            "type": "string",
            "enum": [
              "BRANCH",
              "PLANT",
              "STORE",
              "DEPARTMENT",
              "OTHER"
            ]
          },
          "hasSubledger": {
            "type": "boolean"
          }
        },
        "required": [
          "legalEntityId",
          "code",
          "name"
        ]
      },
...
    }
  }
}